#0: move tile to new file, add hash to desc, debug scan OOM
