+++ tt_metal/impl/buffers/circular_buffer_types.hpp
@@ -14,6 +14,7 @@
 #include "tt_metal/common/tt_backend_api_types.hpp"
 #include "tt_metal/hostdevcommon/common_runtime_address_map.h"
 #include "tt_metal/impl/buffers/buffer.hpp"
+#include "tt_metal/impl/tile/tile.hpp"
 
 namespace tt::tt_metal {
