+++ ttnn/cpp/ttnn/operations/matmul/device/matmul_op_multi_core_reuse_mcast_2d_program_factory.cpp
@@ -520,8 +520,7 @@ operation::ProgramWithCallbacks create_program_mcast_in0_in1(
                 .processor = tt_metal::DataMovementProcessor::RISCV_1,
                 .noc = in0_noc,
                 .compile_args = in0_sender_compile_time_args,
-                .defines = mm_kernel_in0_sender_sharded_defines,
-                .tile_shapes = {in0_tile.get_tile_shape()}});
+                .defines = mm_kernel_in0_sender_sharded_defines});
         if (in0_mcast_cores_without_work_and_not_in_receiver_grid.has_value()) {
             in0_sender_compile_time_args[0] = 0;  // core_has_output_block_work
             in0_sender_compile_time_args[1] = 0;  // core_in_in0_receiver_mcast_grid
@@ -534,8 +533,7 @@ operation::ProgramWithCallbacks create_program_mcast_in0_in1(
                     .processor = tt_metal::DataMovementProcessor::RISCV_1,
                     .noc = in0_noc,
                     .compile_args = in0_sender_compile_time_args,
-                    .defines = mm_kernel_in0_sender_sharded_defines,
-                    .tile_shapes = {in0_tile.get_tile_shape()}});
+                    .defines = mm_kernel_in0_sender_sharded_defines});
         }
     } else {
 
@@ -552,8 +550,7 @@ operation::ProgramWithCallbacks create_program_mcast_in0_in1(
                 .processor = tt_metal::DataMovementProcessor::RISCV_1,
                 .noc = in0_noc,
                 .compile_args = in0_sender_compile_time_args,
-                .defines = mm_kernel_in0_sender_interleaved_defines,
-                .tile_shapes = {in0_tile.get_tile_shape()}});
+                .defines = mm_kernel_in0_sender_interleaved_defines});
     }
 
     auto mm_kernel_in1_sender_writer_id = tt_metal::CreateKernel(
@@ -564,8 +561,7 @@ operation::ProgramWithCallbacks create_program_mcast_in0_in1(
             .processor = tt_metal::DataMovementProcessor::RISCV_0,
             .noc = in1_noc,
             .compile_args = in1_sender_writer_compile_time_args,
-            .defines = mm_kernel_in1_sender_writer_defines,
-            .tile_shapes = {in1_tile.get_tile_shape(), output_tile.get_tile_shape()}});
+            .defines = mm_kernel_in1_sender_writer_defines});
 
     KernelHandle mm_kernel_in1_receiver_writer_id = 0;
     if (in1_receiver.num_cores() > 0) {
@@ -579,8 +575,7 @@ operation::ProgramWithCallbacks create_program_mcast_in0_in1(
                 .processor = tt_metal::DataMovementProcessor::RISCV_0,
                 .noc = in1_noc,
                 .compile_args = in1_receiver_writer_compile_time_args,
-                .defines = mm_kernel_in1_receiver_writer_defines,
-                .tile_shapes = {in1_tile.get_tile_shape(), output_tile.get_tile_shape()}});
+                .defines = mm_kernel_in1_receiver_writer_defines});
     }
 
     KernelHandle mm_kernel_in0_receiver_id = 0;
@@ -593,8 +588,7 @@ operation::ProgramWithCallbacks create_program_mcast_in0_in1(
             tt_metal::DataMovementConfig{
                 .processor = tt_metal::DataMovementProcessor::RISCV_1,
                 .noc = in0_noc,
-                .compile_args = in0_receiver_compile_time_args,
-                .tile_shapes = {in0_tile.get_tile_shape()}});
+                .compile_args = in0_receiver_compile_time_args});
     }
 
     KernelHandle mm_kernel_in1_receiver_writer_other_noc_setup_id = mm_kernel_in1_receiver_writer_id;
@@ -610,8 +604,7 @@ operation::ProgramWithCallbacks create_program_mcast_in0_in1(
                 .processor = tt_metal::DataMovementProcessor::RISCV_0,
                 .noc = in1_split_noc,
                 .compile_args = in1_receiver_writer_compile_time_args,
-                .defines = mm_kernel_in1_receiver_writer_other_noc_setup_defines,
-                .tile_shapes = {in1_tile.get_tile_shape(), output_tile.get_tile_shape()}});
+                .defines = mm_kernel_in1_receiver_writer_other_noc_setup_defines});
 
         mm_kernel_in0_receiver_other_noc_setup_id = tt_metal::CreateKernel(
             program,
@@ -620,8 +613,7 @@ operation::ProgramWithCallbacks create_program_mcast_in0_in1(
             tt_metal::DataMovementConfig{
                 .processor = tt_metal::DataMovementProcessor::RISCV_1,
                 .noc = in0_split_noc,
-                .compile_args = in0_receiver_compile_time_args,
-                .tile_shapes = {in0_tile.get_tile_shape()}});
+                .compile_args = in0_receiver_compile_time_args});
     }
 
     // Compute kernel compile time args
@@ -667,8 +659,7 @@ operation::ProgramWithCallbacks create_program_mcast_in0_in1(
             .fp32_dest_acc_en = fp32_dest_acc_en,
             .math_approx_mode = math_approx_mode,
             .compile_args = compute_kernel_args,
-            .defines = mm_kernel_defines,
-            .tile_shapes = {in0_tile.get_tile_shape(), in1_tile.get_tile_shape(), output_tile.get_tile_shape()}});
+            .defines = mm_kernel_defines});
 
     // Create circular buffers
     uint32_t src0_cb_index = 0;
