+++ ttnn/cpp/ttnn/tensor/tensor.hpp
@@ -18,6 +18,7 @@
 #include "ttnn/common/constants.hpp"
 #include "ttnn/tensor/types.hpp"
 #include "tt_metal/impl/buffers/buffer.hpp"
+#include "tt_metal/impl/tile/tile.hpp"
 #include "tt_metal/impl/device/device.hpp"
 #include "tt_metal/impl/device/mesh_device.hpp"
 #include "tt_metal/tt_stl/reflection.hpp"
